+++ src/check_oldies/check_branches.py
@@ -2,9 +2,10 @@ import argparse
 import os
 import sys
 
-from . import branches
+import check_oldies.branches
+
 from . import configuration
-from . import xunit
+from . import output
 
 
 def get_parser():
@@ -27,11 +28,19 @@ def get_parser():
         ),
     )
     parser.add_argument(
+        "--format",
+        default=output.OutputFormat.TEXT,
+        dest="output_format",
+        help="Output format. Defaults to human-readable text (one result per line).",
+        choices=sorted(output.OutputFormat),
+        type=output.OutputFormat,
+    )
+    parser.add_argument(
         "--max-age",
         type=int,
         help=(
             f"Maximum age in days allowed for a branch, errors otherwise. "
-            f"Defaults to {branches.Config.max_age}."
+            f"Defaults to {check_oldies.branches.Config.max_age}."
         ),
     )
     parser.add_argument(
@@ -41,66 +50,37 @@ def get_parser():
         dest="colorize_errors",
         help="Do not colorize errors. Defaults to colorizing errors in red.",
     )
-    parser.add_argument(
-        "--xunit-file",
-        action="store",
-        help="Path of the xUnit report file to write. Defaults to no xUnit output.",
-    )
     return parser
-
-
-def branch_str(branch):
-    return (
-        f"{branch.author[:30]: <30} - {branch.age: >4} days - {branch.name_and_details}"
-    )
 
 
 def main():
     parser = get_parser()
     config = configuration.get_config(
-        "check-branches", parser, sys.argv[1:], branches.Config
+        "check-branches", parser, sys.argv[1:], check_oldies.branches.Config
     )
     if not configuration.is_git_directory(config.path):
         sys.exit(f'Invalid path: "{config.path}" is not a Git repository.')
 
-    if config.colorize_errors:
-        warn = "\033[91m{}\033[0m".format
-    else:
-        warn = lambda text: text  # pylint: disable=unnecessary-lambda-assignment
+    branches = check_oldies.branches.get_branches(config)
+    branches.sort(key=lambda branch: (branch.author, -branch.age, branch.name))
+    has_old_branches = any(branch for branch in branches if branch.is_old)
 
-    all_branches = branches.get_branches(config)
-
-    out = []
-    uncolorized_out = []
-    for branch in sorted(
-        all_branches, key=lambda branch: (branch.author, -branch.age, branch.name)
-    ):
-        line = branch_str(branch)
-        out.append(warn(line) if branch.is_old else line)
-        uncolorized_out.append(line if branch.is_old else line)
-    has_old_branches = any(branch for branch in all_branches if branch.is_old)
-
-    out = os.linesep.join(out)
+    ok_msg = err_msg = ""
     if has_old_branches:
         err_msg = "NOK: Some branches are too old."
-        print(err_msg)
     else:
-        err_msg = ""
-        print("OK: All branches are fresh.")
-    if out:
-        print(out)
+        ok_msg = "OK: All branches are fresh."
 
-    if config.xunit_file:
-        uncolorized_out = os.linesep.join(uncolorized_out)
-        xunit.create_xunit_file(
-            os.path.abspath(config.xunit_file),
-            "check-branches",
-            "branches",
-            "CheckBranches",
-            err_msg=err_msg,
-            stdout=uncolorized_out,
-            stderr="",
-        )
+    output.printer(
+        branches,
+        config.output_format,
+        ok_message=ok_msg,
+        error_message=err_msg,
+        colorize_errors=config.colorize_errors,
+        xunit_suite_name="check-branches",
+        xunit_case_name="branches",
+        xunit_class_name="CheckBranches",
+    )
 
     sys.exit(os.EX_DATAERR if has_old_branches else os.EX_OK)